requirements.txt
Flask==2.3.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.97.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.27.8
paystack-python==0.1.1
python-jose==3.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
gunicorn==21.2.0
